+++ src/Symfony/Component/Yaml/Tests/ParserTest.php
@@ -813,9 +813,9 @@ EOD;
      * @requires function Symfony\Bridge\PhpUnit\ErrorAssert::assertDeprecationsAreTriggered
      * throws \Symfony\Component\Yaml\Exception\ParseException in 4.0
      */
-    public function testParseExceptionOnDuplicate($input, $duplicate_key)
+    public function testParseExceptionOnDuplicate($input, $duplicateKey, $lineNumber)
     {
-        ErrorAssert::assertDeprecationsAreTriggered(sprintf('Duplicate key "%s" detected whilst parsing YAML. Silent handling of duplicates in YAML is deprecated since version 3.2 and will throw \Symfony\Component\Yaml\Exception\ParseException in 4.0.', $duplicate_key), function () use ($input) {
+        ErrorAssert::assertDeprecationsAreTriggered(sprintf('Duplicate key "%s" detected on line %d whilst parsing YAML. Silent handling of duplicate mapping keys in YAML is deprecated since version 3.2 and will throw \Symfony\Component\Yaml\Exception\ParseException in 4.0.', $duplicateKey, $lineNumber), function () use ($input) {
             Yaml::parse($input);
         });
     }
@@ -827,25 +827,25 @@ EOD;
         $yaml = <<<EOD
 parent: { child: first, child: duplicate }
 EOD;
-        $tests[] = array($yaml, 'child');
+        $tests[] = array($yaml, 'child', 1);
 
         $yaml = <<<EOD
 parent:
   child: first,
   child: duplicate
 EOD;
-        $tests[] = array($yaml, 'child');
+        $tests[] = array($yaml, 'child', 3);
 
         $yaml = <<<EOD
 parent: { child: foo }
 parent: { child: bar }
 EOD;
-        $tests[] = array($yaml, 'parent');
+        $tests[] = array($yaml, 'parent', 2);
 
         $yaml = <<<EOD
 parent: { child_mapping: { value: bar},  child_mapping: { value: bar} }
 EOD;
-        $tests[] = array($yaml, 'child_mapping');
+        $tests[] = array($yaml, 'child_mapping', 1);
 
         $yaml = <<<EOD
 parent:
@@ -854,12 +854,12 @@ parent:
   child_mapping:
     value: bar
 EOD;
-        $tests[] = array($yaml, 'child_mapping');
+        $tests[] = array($yaml, 'child_mapping', 4);
 
         $yaml = <<<EOD
 parent: { child_sequence: ['key1', 'key2', 'key3'],  child_sequence: ['key1', 'key2', 'key3'] }
 EOD;
-        $tests[] = array($yaml, 'child_sequence');
+        $tests[] = array($yaml, 'child_sequence', 1);
 
         $yaml = <<<EOD
 parent:
@@ -872,7 +872,7 @@ parent:
     - key2
     - key3
 EOD;
-        $tests[] = array($yaml, 'child_sequence');
+        $tests[] = array($yaml, 'child_sequence', 6);
 
         return $tests;
     }