+++ src/Symfony/Component/Yaml/CHANGELOG.md
@@ -14,8 +14,8 @@ CHANGELOG
    Yaml::parse('!php/const:PHP_INT_MAX', Yaml::PARSE_CONSTANT);
    ```
 
- * Support for silently ignoring duplicate keys in YAML has been deprecated and
-   will lead to a `ParseException` in Symfony 4.0.
+ * Support for silently ignoring duplicate mapping keys in YAML has been
+   deprecated and will lead to a `ParseException` in Symfony 4.0.
 
 3.1.0
 -----